refact(domain-page): nodes
